#536: Renamed `ExporterFactory` to `AbstractExporterFactory`.
