+++ api/src/main/java/org/itsallcode/openfasttrace/api/exporter/ExporterFactory.java
@@ -1,9 +1,8 @@
 package org.itsallcode.openfasttrace.api.exporter;
-import java.io.*;
+
+import java.io.Writer;
 import java.nio.charset.Charset;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.logging.Logger;
 import java.util.stream.Stream;
 
 import org.itsallcode.openfasttrace.api.core.Newline;
@@ -11,44 +10,16 @@ import org.itsallcode.openfasttrace.api.core.SpecificationItem;
 import org.itsallcode.openfasttrace.api.core.serviceloader.Initializable;
 
 /**
- * Super class for factories producing {@link Exporter}s.
+ * Interface for factories producing {@link Exporter}s.
  */
-public abstract class ExporterFactory implements Initializable<ExporterContext>
+public interface ExporterFactory extends Initializable<ExporterContext>
 {
-    private static final Logger LOG = Logger.getLogger(ExporterFactory.class.getName());
-
-    private final String supportedOutputFormat;
-
-    private ExporterContext context;
-
     /**
-     * Creates a new {@link ExporterFactory}.
-     * 
-     * @param supportedOutputFormat
-     *            the format of the supported output format, e.g.
-     *            {@code "html"}.
-     */
-    protected ExporterFactory(final String supportedOutputFormat)
-    {
-        this.supportedOutputFormat = supportedOutputFormat;
-    }
-
-    @Override
-    public void init(final ExporterContext context)
-    {
-        this.context = context;
-    }
-
-    /**
-     * Get the {@link ExporterContext} set by the {@link #init(ExporterContext)}
-     * method.
+     * Get the {@link ExporterContext}.
      * 
      * @return the {@link ExporterContext}.
      */
-    public ExporterContext getContext()
-    {
-        return this.context;
-    }
+    ExporterContext getContext();
 
     /**
      * Returns {@code true} if this {@link ExporterFactory} supports
@@ -58,10 +29,7 @@ public abstract class ExporterFactory implements Initializable<ExporterContext>
      *            the output type to check.
      * @return {@code true} if the given type is supported for exporting.
      */
-    public boolean supportsFormat(final String format)
-    {
-        return this.supportedOutputFormat.equals(format);
-    }
+    boolean supportsFormat(final String format);
 
     /**
      * Create an exporter that is able to export the given output format.
@@ -78,46 +46,8 @@ public abstract class ExporterFactory implements Initializable<ExporterContext>
      *            the items to export
      * @return an exporter instance
      */
-    public Exporter createExporter(final Path file, final String format, final Charset charset,
-            final Newline newline, final Stream<SpecificationItem> itemStream)
-    {
-        if (!supportsFormat(format))
-        {
-            throw new ExporterException("Output format '" + format + "' not supported for export");
-        }
-        final Writer writer = createWriter(file, charset);
-        return createExporter(writer, itemStream, newline);
-    }
-
-    private static Writer createWriter(final Path file, final Charset charset)
-    {
-        if (file == null)
-        {
-            LOG.finest(() -> "Creating exporter for STDOUT using charset " + charset);
-            return new OutputStreamWriter(getStdOutStream(), charset);
-        }
-        LOG.finest(() -> "Creating exporter for file " + file + " using charset " + charset);
-        return createFileWriter(file, charset);
-    }
-
-    // Using System.out by intention
-    @SuppressWarnings("squid:S106")
-    private static PrintStream getStdOutStream()
-    {
-        return System.out;
-    }
-
-    private static Writer createFileWriter(final Path file, final Charset charset)
-    {
-        try
-        {
-            return Files.newBufferedWriter(file, charset);
-        }
-        catch (final IOException e)
-        {
-            throw new ExporterException("Error creating writer for file " + file, e);
-        }
-    }
+    Exporter createExporter(final Path file, final String format, final Charset charset,
+            final Newline newline, final Stream<SpecificationItem> itemStream);
 
     /**
      * Create an exporter that is able to write to the given file.
@@ -130,6 +60,6 @@ public abstract class ExporterFactory implements Initializable<ExporterContext>
      *            newline format
      * @return an {@link Exporter} instance
      */
-    protected abstract Exporter createExporter(final Writer writer,
-            Stream<SpecificationItem> linkedSpecItemStream, final Newline newline);
+    Exporter createExporter(final Writer writer, Stream<SpecificationItem> linkedSpecItemStream,
+            final Newline newline);
 }
